implementacion de carrito( aun no terminado ), y avance en pruebas unitarias

+++ app/src/main/java/com/example/tiendahuertohogar/data/model/Producto.kt
@@ -12,7 +12,7 @@ data class Producto(
     val codigo: String,
     val categoria: String,
     val nombre: String,
-    val precio: Int,
+    val precio: Double,
     val descripcion: String,
     val personalizable: Boolean,
     val imagenResId: Int = 0
